Initial functionality now achieved, with the test class functional. The approach was changed in several areas. But with the particular addition of BaseArchetype.eventSkillSet being added.
A map of Events and Sets of BaseSkill.java instances, that ensure that, when an event is triggered within ListenerTest.java, it will check if any skills need to be triggered due to that event.

Next few commits will be focused on improving the code's architecture and structure. With particular focus on reducing the need for a lot of event listeners and calls for each individual BaseArchetype.java child class within the plugin.