Progression system has begun by implementing a Progression.java interface, utilised by the ArchetypeExperience.java. This is then tracked a ProgressionListener.java that contains an instance of the CustomExperienceManager.java.
This allows for a system of custom experience orbs to be made as child-classes to ArchetypeExperience.java, that can then be utilised to allow a mechanical way to level up Archetypes through various activities.

Next commit will involve integrating the custom experience and Progression.java system into the BaseArchetype class and existing Archetype package structure.